[FEATURE] Add `Guard` class
The class instances check, whether some object follow the expected assertion.

  class AdultOnly < Assertion::Guard
    attribute :user

    def state
      IsAdult[user.attributes]
    end
  end

The guard instance should be initialized with an object.
Then its `#call` method returns the object itself

  joe = User.new name: "Joe", age: 48
  AdultOnly.new(joe).call
  # => #<User @name="Joe", @age=48>

When the assertion is wrong (the corresponding state invalid),
the guard raises the exception from the assertion:

  ian = User.new name: "Ian", age: 5
  AdultOnly.new(ian).call
  # => #<Assertion::InvalidError @messages=["Ian is a child (age 5)"]>

The guard has a `[]` shortcut to initialize the instance and
immediately call it:

  AdultOnly[joe]
  # => #<User @name="Joe", @age=48>

  AdultOnly[ian]
  # => #<Assertion::InvalidError @messages=["Ian is a child (age 5)"]>

== Additions

* Assertion::Guard